+++ src/gatsby.ts
@@ -1,34 +1,45 @@
 import execa from 'execa'
 import path from 'path'
 
-import addScript from '../utils/add-script'
-import log from '../utils/log'
-import copyDir from '../utils/copyDir'
+import addScript from './lib/addScript'
+import log from './lib/log'
+import copyDir from './lib/copyDir'
 
 const scripts: { [key: string]: string } = {
-  dev: 'next dev',
-  build: 'next build',
-  start: 'next start',
+  build: 'gatsby build',
+  develop: 'gatsby develop --host 0.0.0.0',
+  dev: 'npm run develop',
+  start: 'npm run develop',
 }
 
 const dependencies: string[] = [
   'react',
   'react-dom',
-  'next',
+  'gatsby',
+  // style
   'styled-components',
-  'dotenv',
+  // typescript
+  'gatsby-plugin-typescript',
+  // needed
+  'react-helmet',
+  'gatsby-image',
+  'gatsby-plugin-react-helmet',
+  'gatsby-source-filesystem',
+  'gatsby-transformer-sharp',
+  'gatsby-plugin-sharp',
+  'gatsby-plugin-manifest',
+  'gatsby-plugin-styled-components',
 ]
 
 const devDependencies: string[] = [
   'typescript',
   '@types/react',
-  '@types/node',
   '@types/styled-components',
   '@svgr/webpack',
   'tsconfig-paths-webpack-plugin',
 ]
 
-async function next(name: string) {
+async function gatsby(name: string) {
   const cwd = path.join(process.cwd(), name)
 
   const spinner = log.step('Creating folder and initializing project')
@@ -41,7 +52,7 @@ async function next(name: string) {
   })
 
   log.step('Adding project files')
-  await copyDir(`${path.join(__dirname, '../templates/next')}`, cwd)
+  await copyDir(`${path.join(__dirname, '../templates/gatsby')}`, cwd)
 
   log.step('Installing dependencies')
   await execa('npm', ['i', '--save', ...dependencies], { cwd })
@@ -50,4 +61,4 @@ async function next(name: string) {
   spinner.succeed()
 }
 
-export default next
+export default gatsby
