+++ src/index.ts
@@ -3,12 +3,12 @@ import fs from 'fs'
 import chalk from 'chalk'
 import path from 'path'
 
-import log from './utils/log'
+import gatsby from './gatsby'
+import next from './next'
+import cra from './cra'
+import common from './common'
 
-import gatsby from './lib/gatsby'
-import next from './lib/next'
-import cra from './lib/cra'
-import common from './lib/common'
+import log from './lib/log'
 import overrides from './lib/overrides'
 import initGit from './lib/git'
 import addSrc from './lib/addSrc'
