+++ src/lib/addScript.ts
