+++ src/lib/installMissingDeps.ts
