+++ src/next.ts
@@ -1,45 +1,34 @@
 import execa from 'execa'
 import path from 'path'
 
-import addScript from '../utils/add-script'
-import log from '../utils/log'
-import copyDir from '../utils/copyDir'
+import addScript from './lib/addScript'
+import log from './lib/log'
+import copyDir from './lib/copyDir'
 
 const scripts: { [key: string]: string } = {
-  build: 'gatsby build',
-  develop: 'gatsby develop --host 0.0.0.0',
-  dev: 'npm run develop',
-  start: 'npm run develop',
+  dev: 'next dev',
+  build: 'next build',
+  start: 'next start',
 }
 
 const dependencies: string[] = [
   'react',
   'react-dom',
-  'gatsby',
-  // style
+  'next',
   'styled-components',
-  // typescript
-  'gatsby-plugin-typescript',
-  // needed
-  'react-helmet',
-  'gatsby-image',
-  'gatsby-plugin-react-helmet',
-  'gatsby-source-filesystem',
-  'gatsby-transformer-sharp',
-  'gatsby-plugin-sharp',
-  'gatsby-plugin-manifest',
-  'gatsby-plugin-styled-components',
+  'dotenv',
 ]
 
 const devDependencies: string[] = [
   'typescript',
   '@types/react',
+  '@types/node',
   '@types/styled-components',
   '@svgr/webpack',
   'tsconfig-paths-webpack-plugin',
 ]
 
-async function gatsby(name: string) {
+async function next(name: string) {
   const cwd = path.join(process.cwd(), name)
 
   const spinner = log.step('Creating folder and initializing project')
@@ -52,7 +41,7 @@ async function gatsby(name: string) {
   })
 
   log.step('Adding project files')
-  await copyDir(`${path.join(__dirname, '../templates/gatsby')}`, cwd)
+  await copyDir(`${path.join(__dirname, '../templates/next')}`, cwd)
 
   log.step('Installing dependencies')
   await execa('npm', ['i', '--save', ...dependencies], { cwd })
@@ -61,4 +50,4 @@ async function gatsby(name: string) {
   spinner.succeed()
 }
 
-export default gatsby
+export default next
